reuse test scripts to test images

+++ chamaeleon/test.sh
@@ -1,5 +1,9 @@
-target/debug/chamaeleon &
-pid=$!
+if [[ $1 = "docker" ]]; then
+    docker run -d -p 8000:8000 --name chamaeleon_test chamaeleon:test
+else
+    target/debug/chamaeleon &
+    pid=$!
+fi
 
 sleep 1
 
@@ -8,7 +12,11 @@ curl -X POST http://localhost:8000 \
     -H "Content-Type: text/plain" \
     -d "## Hello *world*!")
 
-kill $pid
+if [[ $1 = "docker" ]]; then
+    docker stop chamaeleon_test -t 3
+else
+    kill $pid
+fi
 
 if [[ $converted != "<h2>Hello <em>world</em>!</h2>" ]]; then
     exit 1
